fix: rename subpath detection
